+++ src/text/juce_String.cpp
@@ -78,7 +78,7 @@ public:
     static juce_wchar* createCopy (const char* const src, const size_t numChars)
     {
         juce_wchar* const dest = createUninitialised (numChars);
-        CharacterFunctions::copy (dest, src, numChars);
+        CharacterFunctions::copy (dest, src, (int) numChars);
         dest [numChars] = 0;
         return dest;
     }
